Adapt tests (#3615)
* tidy up adapt tests

* remove some unnecessary stuff
