+++ packages/kit/.gitignore
@@ -5,5 +5,5 @@
 /client/**/*.d.ts
 /test/**/.svelte-kit
 /test/**/build
-!/src/core/adapt/test/fixtures/*/.svelte-kit
+!/src/core/adapt/fixtures/*/.svelte-kit
 !/test/node_modules
