+++ packages/kit/src/core/adapt/fixtures/basic/static/answer.md
