--- packages/kit/src/core/adapt/test/fixtures/prerender/.svelte-kit/output/server/app.js
@@ -1,12 +1,0 @@
-export class App {
-	render() {
-		return new Response('', {
-			status: 200,
-			headers: {
-				'content-type': 'text/html'
-			}
-		});
-	}
-}
-
-export function override() {}
