--- packages/kit/src/core/adapt/test/fixtures/prerender/.svelte-kit/output/server/manifest.js
@@ -1,1 +1,0 @@
-export const manifest = {};
